+++ peppol-common/src/main/java/no/difi/vefa/peppol/common/model/InstanceIdentifier.java
@@ -1,0 +1,14 @@
+package no.difi.vefa.peppol.common.model;
+
+public class InstanceIdentifier {
+
+    private String value;
+
+    public InstanceIdentifier(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}